interact-with-gcp/requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
google-api-core==2.11.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
oauthlib==3.2.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.14
